% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTMB.R
\name{annotateTMB}
\alias{annotateTMB}
\title{Annotate TMB dataframe with clinical outcome}
\usage{
annotateTMB(TMB_df, ClinicalData)
}
\arguments{
\item{TMB_df}{output from applyTMB()}

\item{ClinicalData}{Simple dataframe with at least two columns, a
Sample column and a ClinicalResponder colum}
}
\value{
annotate dataframe which combines both TMB and clinical responders
}
\description{
Add a new column to the dataframe with the clinical outcome
}
\examples{
# Import datasets
data(TMB_VanAllen)
data("VanAllen_Clinical")
# Annotate TMB with clinical reponse
TMB_clincal_response <- annotateTMB(TMB_df = TMB_VanAllen, ClinicalData = VanAllen_Clinical )
}
