% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilters.R
\name{applyFilters}
\alias{applyFilters}
\title{Apply filters on a list of CollapsedVCF objects}
\usage{
applyFilters(
  vcfs,
  assembly,
  design,
  vaf.cutoff = 0,
  remove.nonexonic = TRUE,
  remove.cancer = FALSE,
  tsList = NULL,
  variantType = c()
)
}
\arguments{
\item{vcfs}{a \code{list} of one or more \code{CollapsedVCF} object(s)}

\item{assembly}{human genome assembly: hg19 or hg38}

\item{design}{a \code{GRanges} object containing WES or panel design}

\item{vaf.cutoff}{minimum value of variant allele frequency accepted}

\item{remove.nonexonic}{logical value 'TRUE' or 'FALSE' indicating whether or
not SNV mapped ouside of exons are to be removed. Default is True}

\item{remove.cancer}{logical value 'TRUE' or 'FALSE' indicating whether or
not to remove cancer variants (variants described in COSMIC and
truncating mutations in tumor suppressors)}

\item{tsList}{path to file containing list of tumor suppressors. If not
provided a list of 1217 tumor suppressors from the TSgene2 database
(<http://bioinfo.mc.vanderbilt.edu/TSGene/>) is used by default.}

\item{variantType}{type of variant to remove. Possible values: synonymous,
nonsynonymous, frameshift, nonsense, not translated or a  combination of them
specified in a \code{character vector}}
}
\value{
Returns a \code{list} of \code{lists}. Each \code{list} include the
following elements: a \code{GRanges}, \code{CollapsedVCF}, \code{data.frame}
object containing variants passing the filter, a \code{charcater string}
describing applied filter (if any), a \code{GRanges} or \code{character vector}
with the sequencing design, a \code{character string} with the name of the
sample.
}
\description{
This function calls the callFilters() function on each element of a list of
CollapsedVCF objects and generates in output a list of lists. Each list
corresponds to one input element, and consists in four elements: an object
containg variants which passed the filter, a character string describing the
applied filter (if any), an object containing the sequencing design and a
character string with the name of the sample.
}
\examples{

## Read vcf
vcf_files <- list(Horizon5="Horizon5_ExamplePanel.vcf", 
                  HorizonFFPEmild="HorizonFFPEmild_ExamplePanel.vcf")
vcf_files <- lapply(vcf_files
                    , function(x) system.file("extdata", x, package = "TMBleR", mustWork = TRUE))
vcfs <- readVcfFiles(vcfFiles = vcf_files, assembly = "hg19")

## Read design
design <- readDesign(system.file("extdata"
, "ExamplePanel_GeneIDs.txt"
, package = "TMBleR"
, mustWork = TRUE)
, assembly = "hg19"
, ids = "entrezgene_id")

## Apply filter to remove known cancer variants using the default tumor
## suppressors list
vcfs_NoCancer <- applyFilters(vcfs = vcfs
, assembly = "hg19"
, design = design
, remove.cancer = TRUE
, tsList = NULL
, variantType = NULL)

## Apply filter to remove synonymous mutations
vcfs_filtered <- applyFilters(vcfs = vcfs,
                             assembly = "hg19",
                             design = design,
                             remove.cancer = FALSE,
                             tsList = NULL,
                             variantType = "synonymous")


}
\author{
Laura Fancello
}
