% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFilters.R
\name{callFilters}
\alias{callFilters}
\title{Filters variants according to input arguments}
\usage{
callFilters(
  vcf,
  assembly,
  design,
  vaf.cutoff,
  remove.cancer,
  remove.nonexonic,
  tsList,
  variantType
)
}
\arguments{
\item{vcf}{\code{CollapsedVCF} object}

\item{assembly}{human genome assembly: hg19 or hg38}

\item{design}{a \code{GRanges} object containing WES or panel design}

\item{vaf.cutoff}{minimum value of variant allele frequency accepted}

\item{remove.cancer}{logical value 'TRUE' or 'FALSE' indicating whether or
not to remove cancer variants, which is variants described in COSMIC or
truncating mutations in tumor suppressors}

\item{remove.nonexonic}{logical value 'TRUE' or 'FALSE' indicating whether or
not to remove SNV mapped in non exonic regions}

\item{tsList}{path to file containing list of tumor suppressors. If not
provided a list of 1217 tumor suppressors from the TSgene2 database
(<https://bioinfo.uth.edu/TSGene/>) is used.}

\item{variantType}{type of variant to remove: synonymous, frameshift or
nonsense}
}
\value{
Returns a \code{list} with the following elements: a \code{GRanges},
\code{CollapsedVCF}, \code{data.frame} object containing variants passing the
filter, a \code{charcater string} describing applied filter (if any), a
\code{GRanges} or \code{character vector} with the sequencing design,.
}
\description{
This function calls different filtering functions on the input vcf object
according to the input arguments and gives in output a list with the three
elements: an object containg variants which passed the filter, a character
string describing the applied filter (if any) and an object containing the
sequencing design
}
\author{
Laura Fancello
}
