% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callTMB.R
\name{callTMB}
\alias{callTMB}
\title{Perform Tumor Mutational Burden (TMB) quantification}
\usage{
callTMB(inputForTMB, genome)
}
\arguments{
\item{inputForTMB}{a \code{list} with the following elements: sample name, 
filter description, design and a \code{CollapsedVCF} or \code{data.frame} 
object containing somatic variants}

\item{genome}{human genome assembly: hg19 or hg38}
}
\value{
Returns a \code{table} with sample name, filter description, size of
the sequenced genomic space and TMB expressed both as total number of
mutations in the sequenced space and number of mutations per megabase
}
\description{
This function counts the total number of mutations in the object provided in
input with variants. It gives in output both the total number of mutations and
the number of mutations per Mb (dividing by the sequenced genomic space,
extracted from the sequencing design).
}
\author{
Laura Fancello
}
