% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareROC.R
\name{compareROC}
\alias{compareROC}
\title{Generate and compare two ROC curves of TMB performance in immunotherapy responder
classification}
\usage{
compareROC(dataset, TMB1, TMB2, method, paired, boot.n)
}
\arguments{
\item{dataset}{a \code{data.frame} containing numeric vectors of TMB values 
and a factor with name 'ClinicalResponse'  and levels 'responder' and 
'nonresponder'}

\item{TMB1}{\code{character string} indicating the \code{data.frame} column
containing TMB values from the first approach}

\item{TMB2}{\code{character string} indicating the \code{data.frame} column
containing TMB values from the second approach}

\item{method}{\code{character string} indicating the method to use: delong,
bootstrap or venkatraman. See \code{\link[pROC]{roc.test}} for more details.}

\item{paired}{a logical indicating whether you want a paired roc.test. If
NULL, the paired status will be auto-detected by are.paired. If TRUE but the
paired status cannot be assessed by are.paired will produce an error.
See \code{\link[pROC]{roc.test}} for more details.}

\item{boot.n}{for method="bootstrap" and method="venkatraman" only: the
number of bootstrap replicates or permutations. Default: 2000.
See \code{\link[pROC]{roc.test}} for more details.}
}
\value{
Returns in standard output the result of the test used to compare the
two ROC curves and a pdf file containg plots of the ROC curves and their
respective AUC, sensitivity and specificity values and the best
threshold for sample classification.
}
\description{
This function takes in input two sets of TMB values for the same samples and
a binary variable representing the clinical response to immunotherapy. It
generates two ROC curves showing performance of the two TMB datasets to
discriminate between responders and nonresponders, as well as the best TMB
cutoff.
}
\examples{
## Compare the ROC curves representing the performance of panel-based and
## WES-based TMB to predict clinical response with the default delong method.

# Set the seed to create reproducible train and test sets in generateROC
set.seed(949)


# Read TMB values and response to immunotherapy
data(Hellman_SimulatedFM1Panel_WES)

# Compare ROCs
compareROC(dataset = Hellman_SimulatedFM1Panel_WES,
           TMB1 = "Panel.NumMuts",
           TMB2 = "WES.NumMuts",
           method = "d",
           paired = TRUE)

}
\seealso{
\code{\link[pROC]{roc.test}}
}
\author{
Laura Fancello
}
