% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTMBdistributions.R
\name{compareTMBdistribution}
\alias{compareTMBdistribution}
\title{Compare TMB distribution in responders and nonresponders}
\usage{
compareTMBdistribution(dataset, TMB)
}
\arguments{
\item{dataset}{a \code{data.frame} object with the following fields:
patient identifier, numeric TMB value and response to
immunotherapy, expressed as "responders" or "nonresponders"}

\item{TMB}{name of the \code{data.frame} column containing TMB values}
}
\value{
Returns the statistics from Wilcoxon test
}
\description{
This function generates applies the unpaired Mann-Whitney test to compare the
 distribution of TMB values between immunotherapy responders and nonresponders.
}
\examples{
## Compare TMB distribution between immunotherapy responders and nonresponders
## using the unpaired Mann-Whitney test

# Read TMB values and response to immunotherapy
data(Hellman_SimulatedFM1Panel_WES)

# Compare TMB distribtions by Wilcoxon test
compareTMBdistribution(dataset = Hellman_SimulatedFM1Panel_WES, TMB = "WES.NumMuts")


}
\author{
Laura Fancello
}
