% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_hg19_demo}
\alias{COSMIC_hg19_demo}
\title{Cosmic coding mutations hg19}
\format{
A data frame with 4740646 rows and 6 variables
\describe{
 \item{CHR}{chromosome name, in UCSC format chrN}
 \item{START}{start genomic coordinate, UCSC format}
 \item{END}{start genomic coordinate, UCSC format}
 \item{STRAND}{strand}
 \item{REF}{reference nucleotide}
 \item{ALT}{alternative nucleotide}
}
}
\source{
From CosmicCodingMuts.vcf file for hg19 downloaded from
https://cancer.sanger.ac.uk/cosmic/download and saved in data-raw as
CosmicCodingMuts_hg19.vcf. The dataset was then subset through random 
stratified sampling to max 10.000 snv per chromosome.
}
\usage{
COSMIC_hg19_demo
}
\description{
A dataset containing coding mutations from COSMIC for genome assembly hg19. 
Since the COSMIC dataset is very large and it requires registration we are
bundleing in the package a very small version, just enough to demo the
package vignette and test suite. Please refer on the documentation 
vignette/HOWTO_Import_external_data.html on how to retrieve and 
use COSMIC data with TMBleR
data("COSMIC_hg19_demo") will load the demo dataset as a "COSMIC_hg19
 object in the global env
}
\keyword{datasets}
