% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_filtered_SNV.R
\name{.fetch_filtered_SNV}
\alias{.fetch_filtered_SNV}
\title{.fetch_filtered_SNV}
\usage{
.fetch_filtered_SNV(original_vcf, filtered_vcf)
}
\arguments{
\item{original_vcf}{Granges object}

\item{filtered_vcf}{Granges object}
}
\value{
vcfcollabpsed object
}
\description{
This is just a helper function that can be used to retrieved the SNV
that have been filtered out
}
