% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_coding_SNV.R
\name{.remove_noncoding_SNV}
\alias{.remove_noncoding_SNV}
\title{.remove_noncoding_SNV}
\usage{
.remove_noncoding_SNV(vcf, assembly)
}
\arguments{
\item{vcf}{VariantAnnotation vcf object}

\item{assembly}{hg19 or hg38}
}
\value{
Granges object
}
\description{
This helper function removes from a VariantAnnotation VCF object the 
SNV mapped in non coding regions
}
