% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamplePaneldesign}
\alias{ExamplePaneldesign}
\title{Dataset containing sequencing design of a panel}
\format{
A GRanges object with 8320 ranges and 2 metadata columns
\describe{
  \item{seqnames}{Chromosome name}
  \item{ranges}{Start-End genomic positions}
  \item{strand}{'+', '-' or '*' for undefined}
  \item{V4}{hgnc gene symbol and start-end cooridnates}
  \item{V6}{hgnc gene symbol and Pool (1 or 2)}
}
}
\source{
In house panel design, provided as example.
design_file_path = system.file( "data-raw"
    , "ExamplePaneldesign.bed"
    , package = "TMBleR"
    , mustWork = TRUE)
    
design <- readDesign(
    filename = design_file_path
    , assembly = "hg19
    , ids = "entrezgene_id")
}
\usage{
ExamplePaneldesign
}
\description{
Sequencing design of the panel used as examples.
}
\keyword{datasets}
