% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleWESdesign}
\alias{ExampleWESdesign}
\title{Dataset containing sequencing design of a WES experiment}
\format{
A GRanges object with 199202 ranges and 0 metadata columns
\describe{
  \item{seqnames}{Chromosome name}
  \item{ranges}{Start-End genomic positions}
  \item{strand}{'+', '-' or '*' for undefined}
}
}
\source{
ss_v6_merged+exome.bed file containing design of an example WES
}
\usage{
ExampleWESdesign
}
\description{
Sequencing design of the WES experiments used as examples.
}
\keyword{datasets}
