% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleWESvcfs}
\alias{ExampleWESvcfs}
\title{Dataset containing example vcf from WES}
\format{
A list of four elements (corresponding to 4 samples). Each element is a CollapsedVCF object.
\describe{
  \item{fixed}{dataframe with REF, ALT, QUAL, FILTER}
  \item{info}{dataframe with data in INFO field of VCF}
  \item{rowRanges}{GRanges object with chromosome, start and end coordinates, strand}
  \item{colData}{dataframe with sample name}
  \item{assays}{Reference class object of class "ShallowSimpleListAssays", describing fields in INFO of VCF}
  \item{elementMetadata}{dataframe}
  \item{NAMES}{NULL}
  \item{metadata}{dataframe with metadata of VCF}
}
}
\source{
in house examples of WES vcfs.
}
\usage{
ExampleWESvcfs
}
\description{
Mutations in VCF format from WES sequencing of 4 samples
}
\keyword{datasets}
