% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByVAF.R
\name{filterByVAF}
\alias{filterByVAF}
\title{Filter variants according to variant allele frequency}
\usage{
filterByVAF(vcf, vaf.cutoff)
}
\arguments{
\item{vcf}{a \code{CollapsedVCF} object containing somatic variants}

\item{vaf.cutoff}{minimum value of variant allele frequency accepted}
}
\value{
Returns a \code{CollapsedVCF} object containing only variants with
variant allele frequency above the cutoff
}
\description{
This function removes from the input variants those with variant allele
frequency (VAF) inferior to the cutoff value provided. The AF field is
required in the input vcf file to apply this function.
}
\author{
Laura Fancello
}
