% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCOSMIC.R
\name{filterCOSMIC}
\alias{filterCOSMIC}
\title{Filter variants described in COSMIC}
\usage{
filterCOSMIC(vcf, assembly)
}
\arguments{
\item{vcf}{a \code{data.frame} object containing variants}

\item{assembly}{human genome assembly to use: "hg19" or "hg38"}
}
\value{
Returns a \code{CollapsedVCF} object with variants not described in COSMIC
}
\description{
This function removes from the input variants those described in COSMIC
}
\author{
Laura Fancello
}
