% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVariantType.R
\name{filterVariantType}
\alias{filterVariantType}
\title{Filter variants by the specified type}
\usage{
filterVariantType(vcf, assembly, variantType)
}
\arguments{
\item{vcf}{a \code{CollapsedVCF} or \code{data.frame} object containing
variants}

\item{assembly}{human genome assembly: "hg19" or "hg38"}

\item{variantType}{type of variant to remove. Possible values are: synonymous,
nonsynonymous, frameshift, nonsense, not translated or a combination of them.}
}
\value{
Returns a \code{CollapsedVCF} or \code{data.frame} object containing
those variants which passed the filter.
}
\description{
This function filters out from the variants of the input vcf file those
corresponding to the type specified in the variantType argument: synonymous,
nonsynonymous, frameshift, nonsense, not translated or a combination of them.
}
\author{
Laura Fancello
}
