% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCOSMIC.R
\name{formatCOSMIC}
\alias{formatCOSMIC}
\title{Read in a COSMIC file and format it for usage in TMBleR}
\usage{
formatCOSMIC(input_file, assembly, output_file = "COSMIC")
}
\arguments{
\item{input_file}{path to CosmicCodingMuts.vcf file downloaded following the vignette instructions}

\item{assembly}{either hg19 or hg38}

\item{output_file}{path to where to save the converted .vcf file into a .rda}
}
\value{
nothing
}
\description{
The objective of this script is to convert a COSMIC .vcf file into a R object
}
