% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBoxplot.R
\name{generateBoxplot}
\alias{generateBoxplot}
\title{Generate boxplots of TMB distribution versus clinical response}
\usage{
generateBoxplot(TMB_clinical)
}
\arguments{
\item{TMB_clinical}{dataframe generated by annotateTMB()}
}
\value{
boxplot
}
\description{
This function generates boxplots representing the distribution of TMB
values between immunotherapy responders and nonresponders.
}
\examples{
#' # Import datasets
data(TMB_VanAllen)
data("VanAllen_Clinical")
# Annotate TMB with clinical reponse
TMB_clincal_response <- annotateTMB(TMB_df = TMB_VanAllen, ClinicalData = VanAllen_Clinical )

# generate plot
generateBoxplot(TMB_clincal_response)
}
