% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateROC.R
\name{generateROC}
\alias{generateROC}
\title{Generate ROC curves to evaluate TMB performance in immunotherapy responder
classification}
\usage{
generateROC(dataset, method = "Youden")
}
\arguments{
\item{dataset}{a \code{data.frame} containing a numeric vector of TMB values 
and a factor with name 'ClinicalResponse'  and levels 'responder' and 
'nonresponder' and a column with TMB values named "TMB_per_Mb".}

\item{method}{\code{character string} indicating the method to identify the
best TMB cutoff: i.e. Youden, MaxSe, MaxSpSe. For more details see also
\code{\link[OptimalCutpoints]{optimal.cutpoints}}}
}
\value{
Returns in output the best TMB cutoff for classification,
AUC with 95% C.I. specificity, sensitivity and positive and negative predictive
values based on the model built on training data. It also returns the
confusion matrix of classification on test data based on the identified TMB
cutoff and a the ROC curve on train and test data.
}
\description{
This function takes in input a data frame with a vector of TMB values and a
factor named ClinicalResponse with only these two levels allowed: 'responder'
and 'nonresponder'. It gives in output a ROC curve describing TMB performance
in responders and nonresponders classification on a training set (random 75%
of input data), the best value of TMB cutoff for classification in this 
training dataset, a confusion matrix describing the classification performance
of this TMB cutoff in a training dataset (25% of input data) and a ROC curve 
describing TMB performance in responders and nonresponders classification in 
a training set.
}
\author{
Laura Fancello
}
