% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hellman_SimulatedFM1Panel_WES}
\alias{Hellman_SimulatedFM1Panel_WES}
\title{Dataset containing tumor mutational burden and clinical data from a published
study}
\format{
A data frame with 74 rows and 4 variables
\describe{
  \item{PatientID}{Patient identifier}
  \item{ClinicalResponse}{Clinical response to immunotherapy; responder or
  nonresponder}
  \item{Panel.NumMuts}{Panel-based tumor mutational burden, in total number of
   mutations}
  \item{WES.NumMuts}{WES-based tumor mutational burden, in total number of
  mutations}
  }
}
\source{
provided as supplementary table by Hellman et al. in "Tumor Mutational
 Burden and Efficacy of Nivolumab Monotherapy and in Combination With Ipilimumab
 in Small-Cell Lung Cancer". Cancer Cell 2018 (PMID: 29731394) and modified to
 add simulated panel TMB
}
\usage{
Hellman_SimulatedFM1Panel_WES
}
\description{
Dataset from Hellman et al., Cancer Cell 2018 containing the original
WES-based TMB value and clinical response from the study and a simulated
panel-based TMB value calculated in silico by subsetting the WES dataset to
only contain genes targeted by a custom panel used as example.
}
\keyword{datasets}
