% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanTumorSuppressors}
\alias{HumanTumorSuppressors}
\title{List of human tumor suppressors}
\format{
character vector with 1217 elements
}
\source{
From the TSGene 2.0 database available at
https://bioinfo.uth.edu/TSGene/. See Zhao et al. "TSGene 2.0: an
updated literature-based knowledgebase for tumor suppressor genes" NAR 2016.
PMID: 26590405
}
\usage{
HumanTumorSuppressors
}
\description{
List of gene ids known to be human tumor suppressors. It is used as default
tsList argument in applyFilters()  if not specified otherwise and the
removeCancer argument is set to TRUE. In such a sceneario, truncating mutations in
these human tumor suppressor genes are not counted for TMB quantification.
This is a corretive strategy usually required for panel-based TMB 
quantification to avoid upward skewing bias due to panel enrichment in cancer 
genes.
}
\keyword{datasets}
