% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputToTMB.R
\name{inputToTMB}
\alias{inputToTMB}
\title{Provide a suitable input to the callTMB function for TMB quantification}
\usage{
inputToTMB(vcf, design, filter)
}
\arguments{
\item{vcf}{a \code{CollapsedVCF} object containing variants}

\item{design}{a \code{GRanges} object containing WES or panel sequencing
design}

\item{filter}{a \code{character string} describing the type of filter
previously applied to input variants}
}
\value{
Returns a list with filter description, a design object
and an object containing variants, as required in input by the
\code{\link{callTMB}} function
}
\description{
This function reads in input an object containing mutations and generates in
output a list with the following elements: filter description,
an object with sequencing design and an object containing variants.
}
\author{
Laura Fancello
}
