% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTMB.R
\name{plotTMB}
\alias{plotTMB}
\title{Plots Tumor Mutational Burden (TMB)}
\usage{
plotTMB(TMB_results, type = "barplot")
}
\arguments{
\item{TMB_results}{a \code{data.frame} with sample, filter, sequencing size,
total number of mutations and number of mutations per megabase.}

\item{type}{a string describing the type of plot to generate: 'barplot', 
'densityplot' or boxplot.}
}
\value{
Returns a pot representing TMB per Mb given in input for of all 
samples and with every applied filter separately
}
\description{
This function reads in input the output of applyTMB() function and plots it
}
\author{
Laura Fancello
}
