% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDesign.R
\name{readDesign}
\alias{readDesign}
\title{Read design file}
\usage{
readDesign(filename, assembly, ids, name = NULL)
}
\arguments{
\item{filename}{name of file describing the sequencing design. It requires
".bed" extension for files containing the genomic coordinates of the targeted
regions or ".txt" extension for files containing the gene symbol of targeted
genes. BED files must be tab-delimited, with no track description line and it
must report at least CHR START and END information.}

\item{assembly}{human genome assembly: hg19 or hg38}

\item{ids}{type of gene identifiers, if .txt file with gene list is provided:
hgnc_symbol or entrezgene_id.}

\item{name}{string indicating the name of the design. if not specified, the filename
without extension will be used instead.}
}
\value{
\code{data.frame} with genomic coordinates or a character vector with
symbols of targeted genes
}
\description{
Read the file describing the design of the sequencing experiment (i.e. the
targeted regions in the sequencing experiment). It can be either a BED file
containing genomic coordinates of the targeted genes or a TXT file containing
symbols for targeted genes.
}
\examples{
#Read in input the panel sequencing design (either a BED file with genomic
#coordinates of the targeted genes or TXT file with targeted gene names).
design_FromTxt <- readDesign(filename = system.file("extdata",
                                 "ExamplePanel_GeneIDs.txt",
                                 package = "TMBleR",
                                 mustWork = TRUE),
                     assembly = "hg19",
                     ids = "entrezgene_id")

design_FromBed <- readDesign(filename = system.file("extdata",
                                "ExamplePaneldesign.bed",
                                 package = "TMBleR",
                                 mustWork = TRUE),
                              assembly = "hg19")


}
\author{
Laura Fancello
}
