% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePanel.R
\name{simulatePanel}
\alias{simulatePanel}
\title{Downsample a whole exome sequencing (WES) dataset to a subset only containing
genes targeted in the panel}
\usage{
simulatePanel(WES, WES.design, panel.design, assembly)
}
\arguments{
\item{WES}{a list generated by applyFilters or applyInputToTMB with variants,
filter, design and sample elements.}

\item{WES.design}{a \code{GRanges} object containing WES sequencing design. 
Used to remove off target mutations}

\item{panel.design}{a \code{GRanges} object containing sequencing design of
the panel tyou want to simuate. Used to subset WES.}

\item{assembly}{human genome assembly: "hg19" or "hg38"}
}
\value{
Returns a \code{GRanges} object containing variants from the
simulated panel, which is only those variants from WES falling in the regions
targeted by the panel
}
\description{
This function generates a subset of a whole exome sequencing dataset which
only includes the genes of a given gene panel. This subset may then be used
to simulate panel-based TMB quantification.
}
\author{
Laura Fancello
}
