% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMB_Horizon}
\alias{TMB_Horizon}
\title{Input to the plotTMB function to be used in tests}
\format{
A data frame with 16 rows and 5 variables
\describe{
  \item{Sample}{factor, sample identifier}
  \item{Filter}{factor, type of filter, if any, applied on input mutations}
  \item{Sequencing_Size}{numeric vector, genomic size targeted by the
  sequencing experiment (WES or Panel), in Mb}
  \item{Tot_Number_Mutations}{numeric vector, tumor mutational burden in the
  targeted genomic region}
  \item{TMB_per_Mb}{numeric vector, global tumor mutational burden expressed
  as number of mutations}
}
}
\source{
generated in house as an example using a custom Panel and toy
examples of VCF files from Horizon cell line dataset
}
\usage{
TMB_Horizon
}
\description{
Table containing estimated tumor mutational burden and metadata (sample,
filter, size of the sequenced genomic space) calculated using the Horizon 
demo dataset
}
\keyword{datasets}
