% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMB_VanAllen}
\alias{TMB_VanAllen}
\title{Input to the plotTMB function to be used in tests}
\format{
A data frame with 16 rows and 5 variables
\describe{
  \item{Sample}{factor, sample identifier}
  \item{Filter}{factor, type of filter, if any, applied on input mutations}
  \item{Sequencing_Size}{numeric vector, genomic size targeted by the
  sequencing experiment (WES or Panel), in Mb}
  \item{Design}{Name of the panel experiment}
  \item{Tot_Number_Mutations}{numeric vector, tumor mutational burden in the
  targeted genomic region}
  \item{TMB_per_Mb}{numeric vector, global tumor mutational burden expressed
  as number of mutations}
}
}
\source{
Dataset from Van Allen EM  et al., Science 2015
}
\usage{
TMB_VanAllen
}
\description{
Table containing estimated tumor mutational burden and metadata (sample,
filter, size of the sequenced genomic space) calculated using 4 patients from 
the VanAllen et al. dataset.
}
\keyword{datasets}
