% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMBs_WES}
\alias{TMBs_WES}
\title{WES-based TMB to use as input to the correlateTMBvalues and plotCorrelation
functions in tests}
\format{
A data frame with 4 rows and 5 variables
 \describe{
  \item{Sample}{factor, sample identifier}
  \item{Filter}{factor, type of filter, if any, applied on input mutations}
  \item{Sequencing_Size}{numeric vector, genomic size targeted by the
  sequencing experiment (WES or Panel), in Mb}
  \item{Tot_Number_Mutations}{numeric vector, tumor mutational burden in the
  targeted genomic region}
  \item{TMB_per_Mb}{numeric vector, global tumor mutational burden expressed
  as number of mutations}
}
}
\source{
generated in house as an example of TMB estimated obtained by WES.
}
\usage{
TMBs_WES
}
\description{
Table comntaining estimated tumor mutational burden and metadata (smaple,
filter, size of the sequenced genomic space) obtained by WES.
}
\keyword{datasets}
