% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VanAllen_Clinical}
\alias{VanAllen_Clinical}
\title{Dataset containing tumor mutational burden and clinical data from a published
study}
\format{
A data frame with 110 rows and 2 variables
\describe{
  \item{Sample}{Patient identifier}
  \item{ClinicalResponse}{Clinical response to immunotherapy; responder or
  nonresponder}
  }
}
\source{
provided as supplementary file S2 by Van Allen EM et al. in 
"Genomic correlates of response to CTLA-4 blockade in metastatic melanoma"
 Science 2015 (PMID: 26359337). Only clinical response and patient variable
 are saved here
 
 # load supplementary file from 
 VanAllen_Clinical <- read.table("TableS2_Revised.xlsx")
 
 VanAllen_Clinical %<>%              
   dplyr::rename("Sample" = patient) %>%  # rename patient column into Sample
   dplyr::mutate(ClinicalResponse = ifelse(VanAllen_Clinical$group == "response"
                                           , "responder", "nonresponder")) %>%
   dplyr::select(Sample, ClinicalResponse)
}
\usage{
VanAllen_Clinical
}
\description{
Dataset from Van Allen EM  et al., Science 2015 containing the detailed 
clinical and genome characteristics of the indivitual patients enrolled 
into the study. The datase includes wherether the patients were classified
as either responder or non-responder.
}
\keyword{datasets}
