% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcfs_all}
\alias{vcfs_all}
\title{Data to use as input to the applyTMB functions in tests}
\format{
A data frame with 4 rows and 5 variables
 \describe{
  \item{sample}{character, sample identifier}
  \item{Filter}{character, type of filter, if any, applied on input mutations}
  \item{design}{GRanges object, regions targeted by the sequencing experiment
  (WES or Panel)}
  \item{variants}{data.frame, describing mutations by chromosome, start and
  end coordinates, REF and ALT alleles}
}
}
\source{
generated in house as an example of TMB estimated obtained by WES.
}
\usage{
vcfs_all
}
\description{
List of lists, , generated by applyFilters or applyInputToTMB, containing
required input to the applyTMB function. Each element of the list corresponds
to a sample and is a list with the following elements: sample, filter,
variants and design.
}
\keyword{datasets}
